+++ pkg/processor/trigger/rtsp/helpers.go
@@ -7,8 +7,19 @@ SPDX-License-Identifier: Apache-2.0
 package rtsp
 
 import (
+	"bytes"
+	"encoding/json"
+	"io"
+	"net/http"
+	"os"
+	"os/exec"
+	"strconv"
 	"sync"
 	"time"
+
+	"github.com/bluenviron/gortsplib/v5/pkg/description"
+	"github.com/bluenviron/gortsplib/v5/pkg/format"
+	"github.com/pion/rtp"
 )
 
 type DataProcessorStream struct {
@@ -21,6 +32,17 @@ type DataProcessorStream struct {
 	newBytes   int
 	output     chan *Event
 	stop       chan struct{}
+}
+
+// MediaPipeline handles decoding RTP packets and optional FFmpeg conversion
+type MediaPipeline struct {
+	depacketizers map[uint8]any
+
+	ffmpegCmd    *exec.Cmd
+	ffmpegStdin  io.WriteCloser
+	ffmpegStdout io.ReadCloser
+
+	mu sync.Mutex
 }
 
 func NewDataProcessorStream(
@@ -106,4 +128,124 @@ func (dp *DataProcessorStream) tryEmit() *Event {
 
 func (dp *DataProcessorStream) Output() <-chan *Event {
 	return dp.output
+}
+
+// NewMediaPipeline creates a pipeline for given formats
+func NewMediaPipeline(medias []*description.Media) (*MediaPipeline, error) {
+	mp := &MediaPipeline{
+		depacketizers: make(map[uint8]any),
+	}
+
+	// create depacketizers for supported formats
+	for _, media := range medias {
+		for _, forma := range media.Formats {
+			var depacketizer any
+			var err error
+
+			switch f := forma.(type) {
+			case *format.MPEG4Audio:
+				depacketizer, err = f.CreateDecoder()
+				// t.Logger.Info("TEST 000 MPEG 4")
+			case *format.MPEG1Audio:
+				depacketizer, err = f.CreateDecoder()
+				// t.Logger.Info("TEST 111 MPEG 1")
+				// add future formats here
+			}
+
+			if err == nil && depacketizer != nil {
+				mp.depacketizers[forma.PayloadType()] = depacketizer
+			}
+		}
+	}
+
+	return mp, nil
+}
+
+// StartFFmpeg launches FFmpeg to convert input to PCM
+func (mp *MediaPipeline) StartFFmpeg(sampleRate int, channels int) error {
+	mp.mu.Lock()
+	defer mp.mu.Unlock()
+
+	mp.ffmpegCmd = exec.Command("ffmpeg",
+		"-loglevel", "warning",
+		// "-f", "mp3",
+		"-i", "pipe:0",
+		"-f", "s16le",
+		"-acodec", "pcm_s16le",
+		"-ac", strconv.Itoa(channels),
+		"-ar", strconv.Itoa(sampleRate),
+		"pipe:1",
+	)
+	var err error
+	mp.ffmpegStdin, err = mp.ffmpegCmd.StdinPipe()
+	if err != nil {
+		return err
+	}
+	mp.ffmpegStdout, err = mp.ffmpegCmd.StdoutPipe()
+	if err != nil {
+		return err
+	}
+	mp.ffmpegCmd.Stderr = os.Stderr
+
+	return mp.ffmpegCmd.Start()
+}
+
+// StopFFmpeg stops FFmpeg process
+func (mp *MediaPipeline) StopFFmpeg() {
+	mp.mu.Lock()
+	defer mp.mu.Unlock()
+
+	if mp.ffmpegStdin != nil {
+		_ = mp.ffmpegStdin.Close()
+	}
+	if mp.ffmpegCmd != nil && mp.ffmpegCmd.ProcessState == nil {
+		_ = mp.ffmpegCmd.Process.Kill()
+		mp.ffmpegCmd.Wait()
+	}
+}
+
+// ProcessRTP decodes an RTP packet and returns payload ready for FFmpeg
+func (mp *MediaPipeline) ProcessRTP(pkt *rtp.Packet, forma format.Format) ([]byte, error) {
+	depacketizer, ok := mp.depacketizers[forma.PayloadType()]
+	if !ok {
+		return pkt.Payload, nil
+	}
+
+	switch d := depacketizer.(type) {
+	case interface {
+		Decode(*rtp.Packet) ([][]byte, error)
+	}:
+		aus, err := d.Decode(pkt)
+		if err != nil {
+			return nil, err
+		}
+		if len(aus) == 0 {
+			return nil, nil
+		}
+		return bytes.Join(aus, nil), nil
+	default:
+		return pkt.Payload, nil
+	}
+}
+
+// postToWebhook forwards processed event to the configured webhook
+func (t *rtspTrigger) postToWebhook(body []byte) {
+	payload := map[string]any{
+		"handler_output": string(body),
+	}
+	jsonPayload, _ := json.Marshal(payload)
+
+	req, err := http.NewRequest("POST", t.webhookURL, bytes.NewBuffer(jsonPayload))
+	if err != nil {
+		t.Logger.WarnWith("Failed to create webhook request", "err", err)
+		return
+	}
+	req.Header.Set("Content-Type", "application/json")
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		t.Logger.WarnWith("Failed to POST webhook request", "err", err)
+		return
+	}
+	defer resp.Body.Close()
 }
